+++ firmware/src/Nano33BLE/src/Nano33BLE.cpp
@@ -6,7 +6,7 @@
 #include <ArduinoJson.h>
 #include <chrono>
 
-#include "PPM/PPMOut2.h"
+#include "PPM/PPMOut.h"
 #include "PPM/PPMIn.h"
 #include "dataparser.h"
 #include "trackersettings.h"
